Try fixing tests
