+++ src/lib/SmallStrings/src/SmallStringsChainRulesCoreExt.jl
@@ -1,0 +1,5 @@
+module SmallStringsChainRulesCoreExt
+using ChainRulesCore: @non_differentiable
+using ..SmallStrings: SmallString
+@non_differentiable SmallString(::Any)
+end
